yt-dlp/requirements.txt
minio==7.2.0
python-dotenv==1.0.0
yt-dlp==2024.3.10
requests==2.31.0

